Generate a proper definition list for the entries
As of 28c94ee0b925e the HTML output for a yaml file was roughly this:
    <p>yaml-node-name <dd><p>extracted comments</p></dd> </p>

A <dd> outside a <dl> isn't rendered correctly, so the output was
just a list of paragraphs with identical formatting.

Fix this by making sure our content is within a definition list, resulting in
the following HTML output:

    <dl><dt>yaml-node-name</dt><dd><p>extracted comments</p></dd></dl>

Note how this is a "freestanding" <dl> now, there is no wrapping <p> but that
could be added if backwards-compatibility requires it.